# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Authentication API",
    "version": "1.0.0",
    "description": "Learning Management System Authentication API with comprehensive user management features",
    "contact": {
      "name": "API Support",
      "email": "support@lms.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Request success status"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          }
        }
      },
...
      "Note": {
        "type": "object",
        "properties": {
          "noteId": {
            "type": "integer",
            "description": "Note ID"
          },
          "title": {
            "type": "string",
            "description": "Note title"
          },
          "content": {
            "type": "string",
            "description": "Note content"
          },
          "noteType": {
            "type": "string",
            "enum": [
              "text",
              "document",
              "link",
              "multimedia"
            ],
            "description": "Type of note"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Note tags"
          },
          "userId": {
            "type": "integer",
            "description": "Owner user ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Group": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "integer",
            "description": "Group ID"
          },
          "groupName": {
            "type": "string",
            "description": "Group name"
          },
          "description": {
            "type": "string",
            "description": "Group description"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "restricted"
            ],
            "description": "Group visibility"
          },
          "createdBy": {
            "type": "integer",
            "description": "Creator user ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      },
      "CreateGroupRequest": {
        "type": "object",
        "required": [
          "groupName"
        ],
        "properties": {
          "groupName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Group name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "Group description"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "restricted"
            ],
            "default": "public",
            "description": "Group visibility"
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/groups": {
      "post": {
        "summary": "Create a new study group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              },
              "example": {
                "name": "JavaScript Study Group",
                "description": "Weekly study sessions for JavaScript fundamentals",
                "isPrivate": false,
                "maxMembers": 20
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "group": {
                          "$ref": "#/components/schemas/Group"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/groups/{id}/join": {
      "post": {
        "summary": "Join a study group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "joinCode": {
                    "type": "string",
                    "description": "Join code for private groups"
                  }
                }
              },
              "example": {
                "joinCode": "ABC123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully joined the group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid join code or group is full"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Group not found"
          },
          "409": {
            "description": "Already a member of this group"
          }
        }
      }
    },
    "/groups/{id}/leave": {
      "post": {
        "summary": "Leave a study group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully left the group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Group owner cannot leave (must transfer ownership first)"
          },
          "404": {
            "description": "Group not found or not a member"
          }
        }
      }
    },
    "/groups/{id}/members": {
      "get": {
        "summary": "Get group members with pagination",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of members per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Group members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GroupMember"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not a member of this group"
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
    "/groups/{id}/notes": {
      "get": {
        "summary": "Get notes shared with the group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of notes per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term for note title/content"
          },
          {
            "in": "query",
            "name": "noteType",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "document",
                "link",
                "multimedia"
              ]
            },
            "description": "Filter by note type"
          }
        ],
        "responses": {
          "200": {
            "description": "Group notes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Note"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not a member of this group"
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
...
  },
...
}